outputDatabase is now in spideyTest
