+++ spideyTest.py
@@ -1,10 +1,7 @@
-import os
-from tinydb import TinyDB, Query
+import os, sqlite3
 
 # contains the code to read the database from spidey.py
-
-# database output function
-# this function will output the database to a .txt file called spider_results.txt
+# function to print scraping results to a .txt file called spider_results.txt 
 # the form of the output is as follows:
 # Page title
 # URL
@@ -14,44 +11,73 @@ from tinydb import TinyDB, Query
 # Child Link2 ... ...
 # ——————————————————————————————
 # next page
+# Note: the first 10 child links are those printed
+# and top 10 most common keywords and their frequencies are printed in descending order
 
-def outputDatabase(databaseFile):
-    db = TinyDB(databaseFile)
-    pageTable = db.table('pages')
+def outputDatabase():
+    conn = sqlite3.connect('spidey.sqlite')
+    cur = conn.cursor()
 
-    # if the file already exists, delete it
-    if os.path.exists('spider_results.txt'):
-        os.remove('spider_results.txt')
+    outputTarget = 'spider_results.txt'
+    if os.path.exists(outputTarget):
+        os.remove(outputTarget)
 
-    for page in pageTable.all():
-        # get data from page
-        title = page['title']
-        url = page['url']
-        lastModified = page['lastModified']
-        size = page['size']
-        childLinks = page['childLinks']
-        keywords = page['sortedKeywords']
-        frequencies = page['sortedFrequencies']
+    # Get all pages from the database
+    cur.execute('SELECT * FROM Page')
+    pages = cur.fetchall()
 
-        # format data
-        output = f"{title}\n{url}\n{lastModified}, {size}\n"
-        for i in range(min(10, len(keywords))):
-            output += f"{keywords[i]} {frequencies[i]};\n"
-        for link in range(min(10, len(childLinks))):
-            output += f"{childLinks[link]}\n" 
-        output += "——————————————————————————————\n"
+    # initialize the pageOutput string
+    pageOutput = ''
 
-        # if output contains any non-utf-8 characters, replace them with a question mark
-        output = output.encode('utf-8', errors='replace').decode('utf-8')
+    # Loop through all pages and print their information to the output file
+    for page in pages:
+        # Print the page title, URL, and last modified date and size
+        pageOutput += f'{page[2]}\n{page[1]}\n{page[5]}, {page[6]} bytes\n'
 
-        # write to file
-        with open('spider_results.txt', 'a', encoding='utf-8') as f:
-            f.write(output)
+        # Get the top 10 most common keywords for this page and their frequencies
+        cur.execute('''
+            SELECT Term.term, TermFrequency.frequency
+            FROM TermFrequency JOIN Term ON TermFrequency.term_id = Term.term_id
+            WHERE page_id = ?
+            ORDER BY frequency DESC
+            LIMIT 10
+        ''', (page[0],))
+        keywords = cur.fetchall()
+
+        # Append the keywords and their frequencies
+        keyword_str = '; '.join([f'{kw[0]} {kw[1]}' for kw in keywords])
+        pageOutput += f'{keyword_str}\n'
+
+        # Get the first 10 child links for this page
+        cur.execute('SELECT child_url FROM Link WHERE page_id = ? LIMIT 10', (page[0],))
+        links = cur.fetchall()
+
+        # Append the child links
+        link_str = '\n'.join([link[0] for link in links])
+        pageOutput += f'{link_str}\n'
+
+        # Append the separator
+        pageOutput += '——————————————————————————————\n'
+
+    # make sure the output is in utf-8 encodable
+    try:
+        pageOutput.encode('utf-8')
+    except UnicodeEncodeError:
+        # Replace unencodable characters with "?"
+        pageOutput = pageOutput.encode('utf-8', 'replace').decode('utf-8')
+
+    # Write the pageOutput string to the output file
+    with open(outputTarget, 'w', encoding='utf-8') as f:
+        f.write(pageOutput)
+    
+    # Close the database connection
+    cur.close()
+    conn.close()
 
 def main():
     # output the database to spider_results.txt, if it exists (if not, print an error message)
-    if os.path.exists('spideydb.json'):
-        outputDatabase('spideydb.json')
+    if os.path.exists('spidey.sqlite'):
+        outputDatabase()
     else:
         print("Database file does not exist. Please run spidey.exe first.")
